Check for the existence of actual entries when initializing `ViewHistory`, instead of looking for a `exists` key
Despite a number of changes to `ViewHistory` over the years, the basic functionality hasn't changed that much. One such strange remnant is that we're relying on a `exists` key, in viewer.js, when checking if `ViewHistory` contains anything useful.

This patch changes things, by simply checking for keys on load. This has the added benefit that there's less data which needs to be stringified and written to storage (`LocalStorage`/`SessionStorage`). Not to mention that having to explicitly write an `exists` entry is easy to forget.